Build our own graph with d3 instead of using the one generated by the shap library
It is more or less equivalent, but this way we can more easily select which features
we want to show, we can avoid CSP problems on Phabricator and we can remove some
information which might be confusing for developers.

With this commit, there are also a couple of fixes to the feature importance text,
a higher limit of features to show (7) and an additional filter for features that
are not easy to explain given the past history.